+++ src/main/java/org/example/walletservice/service/TransactionService.java
@@ -2,6 +2,8 @@ package org.example.walletservice.service;
 
 import org.example.walletservice.model.entity.Player;
 
+import java.math.BigDecimal;
+
 /**
  * Shows the amount of funds on the account, displays the history of transactions and makes a debit/credit.
  */
@@ -14,7 +16,7 @@ public interface TransactionService {
 	 * @param inputPlayerAmount The amount entered by the player for a credit transaction.
 	 * @param transactionToken  Token for the current transaction.
 	 */
-	void credit(Player player, double inputPlayerAmount, String transactionToken);
+	void credit(Player player, BigDecimal inputPlayerAmount, String transactionToken);
 
 	/**
 	 * Debits funds from a player's account.
@@ -23,7 +25,7 @@ public interface TransactionService {
 	 * @param inputPlayerAmount The amount entered by the player for a credit transaction.
 	 * @param transactionToken  Token for the current transaction.
 	 */
-	void debit(Player player, double inputPlayerAmount, String transactionToken);
+	void debit(Player player, BigDecimal inputPlayerAmount, String transactionToken);
 
 	/**
 	 * Gets the transaction history of a player.
